requirements.txt
requests
jsontableschema==0.10.0
aenum
six>=1.5
sqlalchemy
lxml
ofxtools
bcrypt
pyYAML